Fix #106745: Subdivision surface crash with more than 8 UV maps
Before 6c774feba2c9a1eb5834, the maximum number of UV maps was
enforced throughout Blender. Now it has to be enforced in the few places
that actually have the limit. In the future the limit of the subsurf
could be lifted, but that isn't done now to keep the crash fix simple.